!13569 clean warning at 2021128 * cov * debug * clean warning

+++ ops/built-in/fusion_pass/graph_fusion/ai_core/pack_fusion_pass.cc
@@ -111,10 +111,9 @@ Status PackFusionPass::Fusion(ge::ComputeGraph& graph, Mapping& mapping, vector<
     ge::AttrUtils::SetInt(concatBaseDesc, "concat_dim", axis);
 
     ge::NodePtr pack_base_node = graph.AddNode(concatBaseDesc);
-    FUSION_PASS_CHECK(
-        pack_base_node == nullptr,
-        VECTOR_FUSION_INNER_ERR_REPORT(FUSED_OP_TYPE.c_str(), "fusedNode is null, fusion failed."),
-        return PARAM_INVALID);
+    FUSION_PASS_CHECK(pack_base_node == nullptr,
+                      VECTOR_FUSION_INNER_ERR_REPORT(FUSED_OP_TYPE.c_str(), "fusedNode is null, fusion failed."),
+                      return PARAM_INVALID);
     fusionNodes.push_back(pack_base_node);
     ge::AttrUtils::SetInt(pack_base_node->GetOpDesc(), "N", nodes_num);
     for (InDataAnchorPtr inAnchorPtr : fusedNode->GetOutDataAnchor(0)->GetPeerInDataAnchors()) {
@@ -133,7 +132,7 @@ Status PackFusionPass::Fusion(ge::ComputeGraph& graph, Mapping& mapping, vector<
                       VECTOR_FUSION_INNER_ERR_REPORT(FUSED_OP_TYPE.c_str(), "packBaseDesc is null, fusion failed."),
                       return PARAM_INVALID);
 
-    for (size_t c = 0; c < static_cast<int64_t>(max_inputs); c++) {
+    for (size_t c = 0; c < static_cast<size_t>(max_inputs); c++) {
       ge::GeTensorDesc fusedDesc2 = fusedNode->GetOpDesc()->GetInputDesc(c);
       FUSION_PASS_CHECK(packBaseDesc->AddInputDesc(c, fusedDesc2) != SUCCESS,
                         VECTOR_FUSION_INNER_ERR_REPORT(FUSED_OP_TYPE.c_str(), "add packBaseDesc failed."),
@@ -150,10 +149,9 @@ Status PackFusionPass::Fusion(ge::ComputeGraph& graph, Mapping& mapping, vector<
     ge::AttrUtils::GetInt(fusedDesc, "axis", axis2);
     ge::AttrUtils::SetInt(packBaseDesc, "axis", axis2);
     ge::NodePtr pack_base_node1 = graph.AddNode(packBaseDesc);
-    FUSION_PASS_CHECK(
-        pack_base_node1 == nullptr,
-        VECTOR_FUSION_INNER_ERR_REPORT(FUSED_OP_TYPE.c_str(), "fusedNode is null, fusion failed."),
-        return PARAM_INVALID);
+    FUSION_PASS_CHECK(pack_base_node1 == nullptr,
+                      VECTOR_FUSION_INNER_ERR_REPORT(FUSED_OP_TYPE.c_str(), "fusedNode is null, fusion failed."),
+                      return PARAM_INVALID);
     fusionNodes.push_back(pack_base_node1);
     ge::AttrUtils::SetInt(pack_base_node1->GetOpDesc(), "N", max_inputs);
 
@@ -180,10 +178,9 @@ Status PackFusionPass::Fusion(ge::ComputeGraph& graph, Mapping& mapping, vector<
     ge::AttrUtils::GetInt(fusedDesc, "axis", axis3);
     ge::AttrUtils::SetInt(packLastDesc, "axis", axis3);
     ge::NodePtr pack_base_node3 = graph.AddNode(packLastDesc);
-    FUSION_PASS_CHECK(
-        pack_base_node3 == nullptr,
-        VECTOR_FUSION_INNER_ERR_REPORT(FUSED_OP_TYPE.c_str(), "fusedNode is null, fusion failed."),
-        return PARAM_INVALID);
+    FUSION_PASS_CHECK(pack_base_node3 == nullptr,
+                      VECTOR_FUSION_INNER_ERR_REPORT(FUSED_OP_TYPE.c_str(), "fusedNode is null, fusion failed."),
+                      return PARAM_INVALID);
     fusionNodes.push_back(pack_base_node3);
     ge::AttrUtils::SetInt(pack_base_node3->GetOpDesc(), "N", final_clear_node_num);
     for (size_t i = 0; i < nodes_num; i++) {
@@ -193,10 +190,9 @@ Status PackFusionPass::Fusion(ge::ComputeGraph& graph, Mapping& mapping, vector<
         packDesc->SetType("Pack");
 
         ge::NodePtr pack_node = graph.AddNode(packDesc);
-        FUSION_PASS_CHECK(
-            pack_node == nullptr,
-            VECTOR_FUSION_INNER_ERR_REPORT(FUSED_OP_TYPE.c_str(), "fusedNode is null, fusion failed."),
-            return PARAM_INVALID);
+        FUSION_PASS_CHECK(pack_node == nullptr,
+                          VECTOR_FUSION_INNER_ERR_REPORT(FUSED_OP_TYPE.c_str(), "fusedNode is null, fusion failed."),
+                          return PARAM_INVALID);
         fusionNodes.push_back(pack_node);
         ge::AttrUtils::SetInt(pack_node->GetOpDesc(), "N", max_inputs);
 
@@ -256,10 +252,9 @@ Status PackFusionPass::Fusion(ge::ComputeGraph& graph, Mapping& mapping, vector<
         LastPackDesc->SetType("Pack");
 
         ge::NodePtr last_pack_node = graph.AddNode(LastPackDesc);
-        FUSION_PASS_CHECK(
-            last_pack_node == nullptr,
-            VECTOR_FUSION_INNER_ERR_REPORT(FUSED_OP_TYPE.c_str(), "fusionNode is null, fusion failed."),
-            return PARAM_INVALID);
+        FUSION_PASS_CHECK(last_pack_node == nullptr,
+                          VECTOR_FUSION_INNER_ERR_REPORT(FUSED_OP_TYPE.c_str(), "fusionNode is null, fusion failed."),
+                          return PARAM_INVALID);
         fusionNodes.push_back(last_pack_node);
         ge::AttrUtils::SetInt(last_pack_node->GetOpDesc(), "N", final_clear_node_num);
         ge::GeTensorDesc PackOutputTensor_2 = LastPackDesc->GetOutputDesc(0);
